Update jquery.ui.markup.js
Added additional logic for data-ui="dialog" to runt he code when the element is clicked and not on page load.  There would be no reason to dun the dialog code for all elements on initial lode like normal happens.  This could probably be expanded with a attribute with more global options.  Ie: data-ui-load="load/click/other event".

Also a data-ui-target="class  selector" was added to specify the target of a dialog.  In this code the dialog options are on the button triggering it, not the actually dialog html code/element.  This is so that one element can have more then one configuration.